Histogram normalization

+++ OtsuBinarizer.h
@@ -8,9 +8,9 @@ public:
 	static PngImage* binarizeOnGpu(PngImage* imageToBinarize);
 private:
 	OtsuBinarizer();
-	static void calculateHistogram(std::vector<unsigned char>& image, std::vector<unsigned int>& histogram);
-	static int findThreshold(std::vector<unsigned int>& histogram, long int totalPixels);
+	static void calculateHistogram(std::vector<unsigned char>& image, std::vector<double>& histogram);
+	static int findThreshold(std::vector<double>& histogram, long int totalPixels);
 	static PngImage* binarizeImage(PngImage * imageToBinarize, int threshold);
-	static void showHistogram(unsigned int * histogram);
+	static void showHistogram(double* histogram);
 };
